fix seo, rename background
