+++ src/app.html
@@ -5,20 +5,6 @@
 		<link rel="icon" href="%sveltekit.assets%/favicon.png" />
 		<meta name="viewport" content="width=device-width" />
 
-		<!-- Open Graph / Facebook -->
-		<meta property="og:type" content="website" />
-		<meta property="og:url" content="%sveltekit.url%" />
-		<meta property="og:title" content="Riley Smith" />
-		<meta property="og:description" content="Software Developer" />
-		<meta property="og:image" content="%sveltekit.assets%/favicon.png" />
-
-		<!-- Twitter -->
-		<meta property="twitter:card" content="summary_large_image" />
-		<meta property="twitter:url" content="%sveltekit.url%" />
-		<meta property="twitter:title" content="Riley Smith" />
-		<meta property="twitter:description" content="Software Developer" />
-		<meta property="twitter:image" content="%sveltekit.assets%/favicon.png" />
-
 		%sveltekit.head%
 	</head>
 	<body>
