+++ src/routes/+layout.svelte
@@ -1,18 +1,31 @@
 <script>
 	import Header from '$lib/ui/header.svelte';
 	import '../app.postcss';
-	import { fade } from 'svelte/transition';
 	import Sidebar from '$lib/ui/sidebar.svelte';
 	import Navbar from '$lib/ui/navbar.svelte';
-	import GridAndDotBackground from '$lib/ui/GridAndDotBackground.svelte';
+	import GridBg from '$lib/ui/gridbg.svelte';
+	import { page } from '$app/stores';
 	let open = false;
 </script>
 
 <svelte:head>
 	<!-- Primary Meta Tags -->
-	<title>Riley Smith</title>
-	<meta name="title" content="Riley Smith" />
-	<meta name="description" content="Software Developer" />
+	<title>{$page.data.meta.title}</title>
+	<meta name="title" content={$page.data.meta.title} />
+	<meta name="description" content={$page.data.meta.description} />
+	<!-- Open Graph / Facebook -->
+	<meta property="og:type" content="website" />
+	<meta property="og:url" content="%sveltekit.url%" />
+	<meta property="og:title" content={$page.data.meta.title} />
+	<meta property="og:description" content={$page.data.meta.description} />
+	<meta property="og:image" content="%sveltekit.assets%/favicon.png" />
+
+	<!-- Twitter -->
+	<meta property="twitter:card" content="summary_large_image" />
+	<meta property="twitter:url" content="%sveltekit.url%" />
+	<meta property="twitter:title" content={$page.data.meta.title} />
+	<meta property="twitter:description" content={$page.data.meta.description} />
+	<meta property="twitter:image" content="%sveltekit.assets%/favicon.png" />
 </svelte:head>
 
 <div class="absolute w-full h-dvh z-20 overflow-hidden">
@@ -26,4 +39,4 @@
 	</div>
 </div>
 
-<GridAndDotBackground className="h-full w-full" />
+<GridBg className="h-full w-full" />
